Separate CLI code from core Kolibri API. Expose Kolibri API on kolibri.main.
