+++ kolibri/main.py
@@ -1,0 +1,7 @@
+# Add imports for exporting the main entry point
+# functions for Kolibri here, so that env setup
+# has already happened by this point
+from kolibri.plugins.utils import disable_plugin  # noqa E402
+from kolibri.plugins.utils import enable_plugin  # noqa E402
+from kolibri.utils.main import initialize  # noqa E402
+from kolibri.utils.server import start  # noqa E402
